Split test action into two jobs
The GitHub Action for running our tests was doing two things: checking
that the tests passed, and then generating and uploading a test coverage
report.

Unfortunately, our tests are somewhat slow, so running them twice in the
same job takes a few minutes. Split them up so they can be run in
parallel.