Move SQLFluff requirements.txt to subdirectory
Having `requirements.txt` in the root of the project causes DigitalOcean
App Platform to detect the app as needing the Python buildpack. Adding
that buildpack seems to throw away the Node output(?), and so the
migrate step fails to run:

    Error: Cannot find module '/workspace/dist/scripts/migrate.js'

Even if that did not cause a build failure, we do not want this dev/CI
tool to be present in the deployed image.

Move the `requirements.txt` file into a new subdirectory, and configure
dependabot to look at that new subdirectory weekly.

Issue #1356 Deployment failure 2024-12-03