Rename `KeycloakUserId` to `KeycloakId`
Broaden the scope of the KeycloakId to allow it to be a Keycloak UUID
for either a Keycloak user or a Keycloak organization.

The goal is to be able to use the broader `KeycloakId` for permission
grants. One can grant permission to either a user or an organization.
Since either are a UUID in Keycloak, we can use the same field in the
permissions objects.

The following reasoning is included based on feedback that suggested
the reasoning was not obvious. The resulting type `KeycloakId` only
guarantees that the value is a UUID. It is no longer so narrow as a
Keycloak user ID. It technically could be any UUID as far as the type
checker is concerned, but this situation is no different to it being
named `KeycloakUserId`, it previously only checked that the value is a
UUID. Like the previous situation, the utility of not naming the type
as `Uuid` is the name signals intent. We expect these Keycloak IDs to
be actual foreign IDs in the Keycloak instance, whether they be user
IDs or organization IDs. By the same reason, the variables in the code
where we expect the given ID to be specifically a user ID keep their
existing (more specific) names. Likewise, variables that we expect to
be specifically an organization ID should be named as such. Both the
type names and variable names convey the intended scope. Validation
that the UUIDs actually exist in Keycloak is a separate concern that
we would not expect a type checker to perform. We expect to pick the
UUIDs from the JSON Web Token of actual users.

Issue #1291