Add Keycloak organization UUID to data providers
Given that an organization might exist in Keycloak prior to the
introduction of the data provider row here or vice versa, we intend to
support setting the Keycloak organization ID optionally. This change
allows the software to defer linking a PDC Keycloak organization to a
PDC service data provider until both exist.

This commit does not make use of the Keycloak organization UUID but
allows the software to store it and make use of it for data providers.

Issue #1291