Add Keycloak organization UUID to data providers
Given that an organization might exist in Keycloak prior to the
introduction of the data provider row here or vice versa, we intend to
support setting the Keycloak organization ID optionally. This change
allows the software to defer linking a PDC Keycloak organization to a
PDC service data provider until both exist.

This commit does not make use of the Keycloak organization UUID but
allows the software to store it and make use of it for data providers.

Technically it may be possible to get away with not storing this UUID
in the PDC service, instead querying Keycloak via directly as needed.
So far we have been able to take a simpler approach of having the PDC
service look solely at the JSON Web Token (JWT) of actual users. This
commit assumes we follow a similar approach. Organization A person B
may wish to grant permission to organization C, therefore the
organization C data is not present in person B's JWT. We could have a
design where the PDC service queries the PDC Keycloak instance
directly for lists of organizations, however, we'd prefer to keep it
simple where the JWT has what we need, and allow grants of permission
only to actually-used/actually-logged-in users and organizations. It
is also convenient to see "organization C has Keycloak UUID [yada]
while organization E is still NULL," meaning someone from organization
C has used PDC while no one from organization E has used PDC yet. The
assumption I had was that person D (part of organization C) might log
in at some point without any action on permissions at all, but at that
moment the user's JWT can be used to link up organization C to a UUID.
Later, when person B wants to grant permissions to organization C,
then the linkage is available, and the organization C's UUID can be
set in the permissions table. Other means of linking UUIDs are
possible (besides from logged-in-users' JWTs) but the key point is the
separation of permission setting from organization UUID linking.

Issue #1291