Add Keycloak organization UUID to changemakers
Given that an organization might exist in Keycloak prior to the
introduction of the changemaker row here or vice versa, we intend to
support setting the Keycloak organization ID optionally. This change
allows the software to defer linking a PDC Keycloak organization to a
PDC service funder changemaker until both exist.

See previous commit (the one for data providers) for more elaboration.

Issue #1291