Enable SQLFluff layout.indent rule with tabs
Remove the layout.indent rule[1] from the list of excluded rules, and
configure it to use tabs, like the rest of our codebase.

Beyond the type of indentation, there are several options available for
configuring layout[2]; I chose the ones that made the most sense to me.

Since there have been other migrations since the SQLFluff rules have
been enabled, also update the list of SQL files to ignore.

Add the previous commit to `.git-blame-ignore-revs`; see the section
"Ignored revisions" in the README for how this file is used.

[1] https://docs.sqlfluff.com/en/stable/reference/rules.html#rule-layout.indent
[2] https://docs.sqlfluff.com/en/stable/configuration/layout.html