Streamline sqlfluff
We added sqlfluff but left it to developers to set it up properly.  This
commit takes a more opinionated approach / adds python3 as an official
dependency of the project.

The benefit is that sqlfluff is now run seamlessly and does not require
any manual setup from the developer.

Issue #1401 Incorporate venv as the officially supported mechanism for running sqlfluff