Use a generator for BundleLoader operations
All of our bundle loader operations had almost identical logic.  The
only differences were the type of entity returned and the query
parameters.

By using a generator we DRY our code and in doing so we also ensure
consistency around universal parameter locations (e.g. offset and limit
will ALWAYS be at the end, and when we add things like db object or auth
context they can always go in the beginning)

Generally the bundle loader signatures stayed the same except for cases
where authentication was accepted; this is actually just a temporary
change for this commit since we're about to make auth context universal
and it will then be accepted as an AuthContext in those cases again.