Update load operations to use a generator
This is similar to the change made to bundle load operations -- instead
of having a bunch of redundant logic for converting a set of function
parameters into query parameters, and processing the result (possibly
throwing a NotFoundError) we can do this in one place.

This will make it easy to add auth context parameters all at once.