Add `authContext` to item loaders
Similar to what we added for bundle loaders we want to make it possible
for our queries to support permission models, which means our queries
need to consistently be passed auth context information.

For now we don't ever actually pass auth context, but as we start adding
more restrictions to various types of data we will increasingly need to
start passing it.

Issue #1407 Improve db utilities to consistently accept an optional authentication state