Trigger a deployment when workflows succeed
The new `deploy` workflow should trigger a deployment of the newly
built (in the `build` workflow) image to Digital Ocean App Platform
when several previous tasks such as `test` and `build` succeeded.

The `build` task is assumed to have updated the `latest` image tag and
therefore the deployment command called by `curl` here should deploy
the image just created in most cases. There may still be a chance of
interleaving workflow executions causing the unintended image to be
deployed. Until a more robust solution is found, this may have to do.

The situation as of writing is that no new code is delivered to
production so this commit is a step toward getting an automated
deployment working in a test environment and we expect to improve it
to deploy to production.

Issue #1410