Add `authContext` to item creation operations
It may seem strange to have an auth context when creating, but the
create operations do return data and that data may contain context that
should be filtered by the database to reflect the user's authorization
level.

This does not actually adjust queries, but it does expose all creation
queries to the auth context pararamters.

Issue #1407 Improve db utilities to consistently accept an optional authentication state