Include production deployment in deploy script
Deploy to production when the call to deploy to test succeeds. Without
this commit, the deployment actions only target the test environment.

While this deployment pipeline does not (yet) validate that the test
environment is working after deployment to test, it at least validates
that the command to deploy succeeded via `needs: deploy-to-test-env`.

Issue #1410