Let changemaker Keycloak organization ID be set
Without this commit, there would be no way for the PDC service code to
set a changemaker's Keycloak organization UUID after initial creation.
With this commit a function is available to set it. While the
capability is not yet exposed to the HTTP API, the first use of the
function is expected to be within the service, perhaps in middleware,
and not yet exposed to users. Another reason to defer exposure in the
HTTP API is to figure out whether we want to use `PUT` or `PATCH` for
this purpose. Unlike the funder and data provider entities,
changemaker has a synthetic key of `id` due to a compound natural key
of `taxId` plus `name`.

Issue #1291