Use generator for updateBaseField operator
We're updating our db operation functions to use generators.  The
generator no longer treats an empty DB response as a "not found" but
rather it is considered a generic / unexpected Error. Instead we should
assert that the record exists before attempting to update it.

Eventually this will be replaced with an upsert once we change base
fields to use a natural key.