Return 404 when changemaker ID not present
Add a test to ensure 404 comes back on a non-existent changemaker.

On a `PATCH /changemakers/:changemakerId`, the generated update
function returns no rows when the changemaker ID does not exist. This
differs from the situation on `PUT` where the ID would be created. To
keep the existing generator with minimal disruption to other classes,
use a new marker `Error` class `NoDataReturnedError` such that the
handler for the `PATCH` can return a 404 instead of 500.

Issue #1291 Keycloak groups to assist with PDC object permissions