Add ability to link fiscal sponsors to changemaker
A fiscal sponsor in PDC terms is another changemaker. To add a fiscal
sponsor to a changemaker, call the new endpoint `PUT
/changemakers/{changemakerId}/fiscalSponsors/{fiscalSponsorChangemakerId}`
To remove the relationship, call the corresponding `DELETE` endpoint.

The `GET /changemakers/{changemakerId}` endpoint now returns a list of
fiscal sponsors. The list is composed of shallow changemakers to avoid
a technical issue of potential infinite recursion. So the caller will
not see the `fields` or `fiscalSponsors` within the `fiscalSponsors`
returned by `GET /changemakers/{changemakerId}`. To get the detailed
data on a changemaker, simply call the same `GET` with the sponsor ID.

There is some ambiguity about whether this explicitly modeled
fiscal sponsorship relationship in PDC is intended to be only current
relationships or both current and past relationships. Assume for now
that it is both current and past, reserving `DELETE` for mistaken data
entry.

Circular sponsorship graphs are supported but not direct
self-sponsorship cycles.

For the purpose of integration with fine-grained permissioning, model
sponsors at the API level as attributes of grant-seekers. This is
opposed to using, for example, a separate `/fiscalSponsorships`
endpoint or exposing only `fiscalSponsees`. In other words, if one has
modify permissions on the changemaker, that should naturally extend to
modifying the sponsor relationship to that changemaker. By modeling
it this way in the API it should make the permissions implementation
relatively straightforward. We may wish to revise this in future to
expose `fiscalSponsees` and/or related `PUT` and `DELETE` verbs on
`/changemakers/:changemakerId/fiscalSponsees/:fiscalSponseeChangemakerId`
but as a first implementation give the changemaker control rather than
the sponsor.

Issue #1485 Support fiscal sponsor relationships between changemakers