Store ephemeral user group associations
Users are associated with user groups via the "organizations" attribute
of their authentication JWT.  We're going to want our queries and
authentication checks to account for those group associations, but we
don't want to have to constantly pass an unpredictably long list of
associations every time we make a query.

This entity, and the middleware that populates it, allows a given API
call to have those associations accounted for in a given HTTP request
context.  The table is labeled "ephemeral" to try to convey the fact
that the data it contains should not be considered accurate outside of
that context.

The implementation of the middleware right now is intentionally simple
and does not attempt to be efficient.  For example, it does not attempt
to re-use any past data in the table (e.g. associations that may have
existed from a past transaction and wouldn't necessarily have to be
re-created).

We may find that having O(n) insert queries for every authenticated API
call is too slow; if that ends up becoming a problem we may determine
that our model for using KeyCloak as the group management and JWTs as
the communication mechanism isn't correct.

These associations are not actually used when making auth decisions yet.
This commit is, however, a prerequisite for that feature.

Issue #Create user-group associations "cache" middleware