Associate funders with opportunities
Opportunities must be associated with a funder going forward; this will
play an important role for our permissions implementation, but it is
also just useful metadata for various use cases of the PDC (knowing who
is funding a given opportunity is helpful).

Since bulk uploads create opportunities we needed to add a funder
attribute to bulk uploads as well.

Issue #1498 Add a funder attribute to opportunities