Add permission check to GET /applicationForms
We should only return application forms that the user is allowed to view

Because this check is done at the query level we need to update some
other places where the user now needs "view" access to be able to
perform their activities.

In future we may decide that `edit` access should also grant `view`
access.

Issue #1406 Utilize permissions