Add unique constraint on source label and fkeys
This commit adds unique constraints on a sources label and it's
potential foreign keys to other entities (organization, dataProvider,
funder). Notably, it does allow for two sources to exist with the same
label so long as they are related to a distinct entity.

Issue #1455 Add Unique Constraint on Name and Foreign Key for sources