Upgrade to Express v5
Express 5 now supports promise-based handlers, which will allow us to
flatten our handlers and middleware without using an inline anonymous
async function.

This update does reveal an issue with the supertest type definitions,
and so we have to disable a linting rule in our tests until that gets
corrected.  I opened a PR for their consideration which should address
the issue [1].

Once this is merged we should be able to improve our handlers and
middleware to consistently use async functions.

Issue #1493 Upgrade to Express 5

[1] https://github.com/DefinitelyTyped/DefinitelyTyped/pull/72237