Ensure transaction is complete before responding
There was a bug in the implementation of the `postProposalVersion`
handler where the http response was sent inside of a transaction.  This
could have meant an improper success result in the event that something
went wrong with the transaction closure.  It also meant that there would
be a brief window between the response return value and the database
actually being updated.

This was resulting in intermittently failing tests.

Issue #1550 POST /proposalVersions provides a response before its transaction is committed