Use auth context agent for `createdBy` population
Several of our entities have a `createdBy` attribute.  We used to have
the middleware populate this value using the auth context, but now that
our queries are aware of the auth context values directly we should
make the query responsible for ensuring appropriate integrity.

This is similar to how the query is responsible for populating
`createdAt`.

This change looks very large because a lot of our tests now need to be
updated to use an auth context when setting up fixtures.

Issue #1479 Use auth context for createdBy population