Add Currency Validation
This commit adds a currency validation to our existing base field type
validation, introducing the 'validator' package[1] which validates
ISO_4217 Currency Codes.  It it is an opinionated validation per the
specifications in Issue #299, where a valid currency is in the form
'{amount} {code}', where amount is a number without any digit grouping
separators, uses a period as the decimal separator (with a required two
decimal values), and code is a 3-letter ISO 4217 currency code. Also,
since the validator package covers mobile phone numbers as well, the
'phone' package was removed so we only have to use the one external
validation library.

[1] https://github.com/validatorjs/validator.js