Enable `operation-4xx-response` lint rule
I did *not* go through to ensure the actual response lists are accurate
overall -- this work will need to be part of a larger refactor.

Note that this also starts the pattern of re-using responses when
possible (e.g. there is no need to re-define 404 every time it's used.);
continuing that work is part of issue #1178.

Issue #1177 Switch our openAPI linter to @redocly/cli