Decouple api docs from code
We were loading some defaults from the api documentation directly; this
was a fine concept at the time, but that coupling makes things like
refactoring the api documentation more complicated.

Realistically we *regularly* need to make sure our api documentation is
in line with implementation.  For instance, when updating a type we need
to remember to update the documentation for that type.  Defaults having
similar considerations is appropriate.

Issue #1178 Break our openAPI json file into reusable pieces rather than a single monolith