Alter basefields to use short_code as primary key
This monstrous refactor commit adds the database migration to change the
primary key of basefields from a generated id to a 'natural' key, in
this case the short code, and refactors all relevant code affected by
this.  Furthermore, it updates the fkey constraints on
base_field_localizations and application_form_fields to accomodate this
new change.

Issue #1141 Use Natural Key as Primary Key for Basefields

Co-authored-by: Harry Minsky <harry.minsky@gmail.com>
Co-authored-by: Daniel Schultz <slifty@gmail.com>