Use triggers on DB tables to log data changes
Database triggers provide a low-level audit trail that is available
regardless of whether a data change was made by the application or by
a user manually logged into the database, for example, via `psql`.

This change adds triggers to existing tables that log data changes to
a new table called `audit_logs`.

The Postgres wiki has an existing example script called
`audit-trigger` that has been extended or modified by a variety of
folks. This audit trigger script formed the basis of this change,
specifically the version shown in the second line of the migration
file named `0055-create-audit-log.sql`. Several adaptations have been
made, namely:
* Included the license directly in the file.
* Reformatted to satisfy SQLfluff.
* Removed the separate schema as it does not serve PDC well. Instead,
* renamed the tables and triggers to include `audit` for clarity.
* Added calls to the `audit_table` for existing tables in migration.
* Removed extraneous wrapper functions.
* Removed the assumption of an integer PK named `id`.
* Log calls to `audit_table` and `deaudit_table` as action 'A', 'S'.
* Added the session/connection/pid.

To begin auditing a table (a new one, for example), call the
`audit_table` function with the table name. To stop auditing a table,
call the `deaudit_table` with the table name. To see which tables are
currently being audited, `select * from audited_tables`.

Issue #1559 Implement audit logging