Remove update operations for baseFields
This commit removes update operations for baseFields and the
corresponding PUT basefields route, in preperation for the primary key
refactor for baseFields.  We've decided that since we are changing the
primary key for a basefield from a generated numerical id to the natural
key shortCode, it no longer makes sense from a RESTful perspective to
allow for updating the baseField by ID, as the ID could be changed by
the update and result in a loss of reference.

Issue #1141 Use Natural Key as Primary key for Basefields