Specify UTC on all connections
We should have our service return timestamps as UTC rather than whatever
the host has been configured with.  The absolute times are all the same,
so this is mostly relevant for tests but consistency in output is
useful.

Note that the service will still accept timestamps in any absolute time.