Add Funder Collaboratives
This commit introduces the FunderCollaborativeMember entity, which
represents a one-to-one relationship between a funder and a funder
collaborative. A manager of a FunderCollaborative is granted the same
permissions they have on their Funder entity for the collaborative
funder. Collaborative Funders are distinguished by the new
'is_collaborative' field.

Issue #1639 Create entity for Funder Collaboratives