Add `sensitivityCategory` to `BaseField`
This category allows a PDC administrator to specify the sensitivity of
data associated with a base field.  We will be able to use this to
make some data publicly available (e.g. for changemaker names) in
addition to actually disallowing the storage of certain data in the PDC.

The mechanisms for actually exposing public data and purging forbidden
data will be implemented in future commits.

Issue #1621 Add the concept of "too sensitive to store data" to base fields