Disallow insertion of forbidden ProposalFieldValues
If the BaseField associated with an ApplicationFormField is marked as
forbidden we do not allow new data to be inserted for that field.
This commit enforces that requirement and returns a 400 error if a user
attempts to create a proposal version that violates the constraint.

Issue #1621 Add the concept of "too sensitive to store data" to base fields