Use a transaction for application form creation
The `POST /applicationForms` handler created multiple entities, but if
the creation of the actual form fields failed the application form would
still exist but in a partial state.  This is incorrect, and the whole
thing should be part of a transaction so that if any step fails there is
no change in data state.