Omit ApplicationFormFields associated with FORBIDDEN BaseFields
If a `BaseField` has a `FORBIDDEN` `sensitivityClassification` then any
existing `ApplicationFormFields` that use that `BaseField` should not be
returned when loading `ApplicationForms`.

Eventually there will also be a mechanism to purge those forbidden form
fields, but for now we will merely stop exposing them.

Issue #1621 Add the concept of "too sensitive to store data" to base fields