Prevent json encoding of forbidden proposal_field_values
We never want to return `FORBIDDEN` ProposalFieldValues. The primary way
to accomplish that is to ensure our queries appropriately condition
against selecting those values in the first place.

This commit is a "belt and suspenders" approach which is only taken on
because of the level of sensitivity that this date represents.  If
a query attempts to select a forbidden ProposalFieldValue it will result
in a database exception.

This doesn't completely prevent access (e.g. if a query accesses field
values without creating an entire row), but it is still a layer of
defense.

Issue #1621 Add the concept of "too sensitive to store data" to base fields