Remove airbnb eslint configuration
Gone are the days of the airbnb eslint configuration being the best in
show!  It's a good set of rules but unfortunately the project is no
longer being maintained and is forcing us to stay pinned to
a now-deprecated version of ESLint.

I added one rule manually since otherwise we saw lint violations related
to arrow functions being required to have an `await` in order to be
allowed to be `async` which seems silly to me.

Issue #1699 Move from airbnb eslint configuration