Upgrade to ESLint 9
ESLint 9 has a brand new way of defining configuration, so this upgrade
comes along side a translation of our config file to the New And Only
Way to configure eslint [1]

I tried my best to keep the rules as close as possible to what we had
before but a few things did change as part of the move to the latest
versions of various rulesets.

A happy byproduct of this commit is we can now re-enable dependabot for
major versions of various dev dependencies that we had frozen due to
incompatibility.

Issue #1699 Move from airbnb eslint configuration
Issue #978 Adjust dependabot settings for ESLint 9

[1] https://eslint.org/docs/latest/use/configure/configuration-files