Set up `love` linting rules
We dropped airbnb because it wasn't supported, but look at this amazing
set of rules! In addition to being actively maintained these rules also
appear more robust (as evidenced by the number of rules we needed to
disable after using the ruleset).

By keeping those rules disabled we are set up to be able to go through
each and either enable it or move the exception to the primary block
where rules are manually reconfigured.

Issue #1699 Move from airbnb eslint configuration