Enable `@typescript-eslint/no-magic-numbers`
It's nice to not use magic numbers; 0 and 1 are really awkward though
and often make semantic sense on their own.

I tried to lean into the spirit of things and gave labels to a few
numbers that might veer too far into the "this is silly to name" arena
(particularly around schema definitions).  That said, I think this rule
is generally a good one and so I'm wary of adding global exceptions even
for 0 / 1 / etc.

Issue #1699 Move from airbnb eslint configuration