Enable `@typescript-eslint/no-magic-numbers`
It's nice to avoid magic numbers!  I did add the exception for numbers
used in object definitions, since in those cases the labels often
provide enough context for the numbers to have semantic meaning.

For example: { min: 0 } is pretty clear when defined in the context of
a jsonSchema, and it felt quite silly to need to define `MIN_FOO_VALUE = 0`
for use in that one location.

Issue #1699 Move from airbnb eslint configuration