Add BaseField category
We're ditching the concept of BaseField `scope` and replacing it with
the more comprehensive concept of BaseField `category`.

This involves the renaming of the `proposal` scope to `project` and the
addition of several more specific types of project data.  The concept of
`organization` data is still there in the category paradigm, and that is
used when decorating ChangeMakers with fields in the same way it was
when we called it `scope`.

Issue #1669 Add the concept of "BaseField Category"