Enable `@typescript-eslint/no-unsafe-type-assertion`
This is a huge one, and also incredibly important for ensuring type
safety.  Without this rule we were allowed to override typescript and
force tsc to believe things about variable data that weren't necessarily
true or enforced.

The most uncomfortable aspect of enabling this rule is in jest's
unfortunate typing of asymmetric matchers (which return `any` instead of
`unknown` and thus trigger other lint complaints related to unsafe use
of `any`).  To get around this we define our own shims for those
matchers which cast the result as unknown.

The other places where this rule has to be overrode are when using
Newtypes and when mocking a subset of a given object in our tests.
These are both compartmentalized in type definitions / mock utility
files so I think we're OK.

Issue #1699 Move from airbnb eslint configuration