Require `keycloakOrganizationId` attributes
We generally do not want to allow optional attributes on `POST` --
either something is read only, or we require the user to be explicit
about what value they intend the entity to hold.

If a value allows `undefined` then the attribute does not have to exist
on the entity at all; if a value can hold `null` then the attribute must
exist, but it can hold a value of `null`.