Enable `eslint-comments/require-description`
For the most part we already *did* have descriptions for when a linting
rule was disabled, but those descriptions were not formatted in a way
that this rule detected.

This change uses the documented format for eslint descriptions.

In one case (Writable) the description is not sufficient so I created
a new issue [1] to capture the work for resolving those.

Issue #1699 Move from airbnb eslint configuration

[1] https://github.com/PhilanthropyDataCommons/service/issues/1730