Add Instructions field to ApplicationFormField
This commit adds the new 'instructions' field to the
ApplicationFormField entity, which is intented to provide clarification
and explicit instruction for a user to fill in a given form
field. Furthermore, it also gives a record of the instructions a user
was operating under at the time of filling out the field, which should
help us interpret user input that varies widely from what we're
expecting.

Issue #1670 Add the concept of instructions to ApplicationFormField