Document the `Source` entity
A data source could be thought of as the changemaker, data provider,
or funder. The PDC allows one more level of grouping beyond those
entities: a Source with a label. It is expected to be not ultimate but
last-hop. Without this commit the intent of the Source entity is not
clear but with this commit it is a bit clearer.

Fixes #1733 Document the difference between Source label and Source
    entity

Fixes #1251 Document that "Source" is "last hop" and not "original"