Add Permissions Checks
This commit integrates funderCollaborativeMembers into our permissions
check, by way of the has_funder_permission.sql and user_to_json.sql
initialization functions. Now, in addition to checking explicit funder
permissions for users (and ephemeral group permissions), we are checking
if the user has permissions on any funders that are members of a
collaborative, and appending those transitive permissions to the user
entity in the application.

Issue #1639 Add Funder Collaboratives