Rename `organization:*` to `organizations`
This change reflects a recent change in PDC Keycloak configuration.

When trying organizations for the first time, there was a workaround
to get `organizations` to show in the JWT by using a new Client scope.
Afterwards, the colon character (`:`) gained new and special meaning
in releases of Keycloak. To maintain `organizations` in the JWT, the
name of the Client scope is now `organizations`.

Issue https://github.com/PhilanthropyDataCommons/auth/issues/57
Issue https://github.com/keycloak/keycloak/issues/40670