Add Funder Collaborative Invitations
This commit adds in the concept of 'Invitations' to funder
collaboratives, which will allow for funders with the requisite
permissions to invite other funders to be collaboraters on collaborative
funders. Invitations are to be responded to by users via a PATCH
request, wherein they will have their status updated from pending to
either rejected or accepted. On invitation acceptance, a
funderCollaborativeMember entity is then created, linking the user to
the collaborative. As with the previous commit, permissioning checks
themselves have not been implemented, and will be implemented
subsequently in a future PR.

Issue #1639 Create entity for funder collaboratives