Add Funder Collaborative Invitations
This commit adds in the concept of 'Invitations' to funder
collaboratives, which allow for funders with the requisite permissions
to invite other funders to be collaboraters on collaborative
funders. The model is such that a system administrator can create a
funder-collaborative-member entity between collaborative funder A and
non-collaborative funder B, which gives a user with permissions on A the
same permissions on B (notably, it is a one way relationship, B does not
have management permissions on A). Once A is a collaborator on B, it has
the permission to invite another collaborator C to join B as a
collaborator. If C accepts the invitation, a funderCollaborativeMember
entity is created, and C now has permissions on B (but not A).

Issue #1639 Create entity for funder collaboratives