Add `opportunity` permissions
This is a new suite of permissions that will allow various types of
access to `Opportunity` entities.  The permissions do not currently
do anything but they can now be assigned and removed by system
administrators to various users and user groups.

This is the first time we have a permission set that is distinct from
other entity permission sets and exposes the fact that the overlap
between the Funder / Data Provider / Changemaker permissions is
coincidential.  We should likely remove the `Permission` enum / api
schema type and replace it with more specific `FooPermission` types in
a future improvement.

Issue #1773 Support OpportunityPermission grants